apply PR feedback
