+++ servicex_app/servicex_app/models.py
@@ -356,6 +356,7 @@ class TransformationResult(db.Model):
     total_events = db.Column(db.BigInteger, nullable=True)
     total_bytes = db.Column(db.BigInteger, nullable=True)
     avg_rate = db.Column(db.Float, nullable=True)
+    created_at = db.Column(DateTime, default=func.now())
 
     __table_args__ = (
         db.UniqueConstraint('file_id', 'request_id', name='uix_file_request'),
@@ -376,7 +377,8 @@ class TransformationResult(db.Model):
             'transform_time': x.transform_time,
             'total-events': x.total_events,
             'total-bytes': x.total_bytes,
-            'avg-rate': x.avg_rate
+            'avg-rate': x.avg_rate,
+            'created_at': x.created_at.isoformat() if x.created_at else None,
         }
 
     def save_to_db(self):
